relate planet orbit to star with parent-child relationship
